Add base tabs fragment
