+++ EmbeddedSocialClient/sdk/src/main/java/com/microsoft/embeddedsocial/ui/fragment/PopularFeedFragmentTabs.java
@@ -1,71 +1,25 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See LICENSE in the project root for license information.
+ */
+
 package com.microsoft.embeddedsocial.ui.fragment;
 
 import com.microsoft.embeddedsocial.base.function.Producer;
 import com.microsoft.embeddedsocial.base.utils.EnumUtils;
-import com.microsoft.embeddedsocial.base.view.SlidingTabLayout;
 import com.microsoft.embeddedsocial.data.model.TopicFeedType;
 import com.microsoft.embeddedsocial.sdk.R;
 import com.microsoft.embeddedsocial.service.IntentExtras;
-import com.microsoft.embeddedsocial.ui.fragment.base.BaseFragment;
+import com.microsoft.embeddedsocial.ui.fragment.base.BaseTabsFragment;
 import com.microsoft.embeddedsocial.ui.util.SimplePagerAdapter;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v4.content.ContextCompat;
 import android.support.v4.view.PagerAdapter;
-import android.support.v4.view.ViewPager;
-import android.view.View;
 
-public class PopularFeedFragmentTabs extends BaseFragment {
-    ViewPager viewPager;
-    SlidingTabLayout slidingTabLayout;
-
-    private final ViewPager.OnPageChangeListener onPageChangeListener = new ViewPager.OnPageChangeListener() {
-        @Override
-        public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
-        }
-
-        @Override
-        public void onPageSelected(int position) {
-            PopularFeedFragmentTabs.this.onPageSelected(position);
-        }
-
-        @Override
-        public void onPageScrollStateChanged(int state) {
-        }
-    };
-
-    public PopularFeedFragmentTabs() {
-        addThemeToMerge(R.style.EmbeddedSocialSdkThemeOverlayContentFragment);
-        addThemeToMerge(R.style.EmbeddedSocialSdkAppTheme_LightBase);
-    }
-
-    @Override
-    protected int getLayoutId() {
-        return R.layout.es_fragment_tabs;
-    }
-
-
-    @Override
-    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        viewPager = view.findViewById(R.id.es_viewpager);
-
-        PagerAdapter adapter = createPagerAdapter();
-        viewPager.setAdapter(adapter);
-        viewPager.setOffscreenPageLimit(adapter.getCount());
-
-        slidingTabLayout = view.findViewById(R.id.es_slidingTabs);
-        slidingTabLayout.setViewPager(viewPager);
-
-        slidingTabLayout.setSelectedIndicatorColors(ContextCompat.getColor(getContext(), R.color.es_tab_bottom_line));
-
-        viewPager.addOnPageChangeListener(onPageChangeListener);
-    }
-
+public class PopularFeedFragmentTabs extends BaseTabsFragment {
     protected PagerAdapter createPagerAdapter() {
-        return new SimplePagerAdapter(getContext(), getActivity().getSupportFragmentManager(),
+        return new SimplePagerAdapter(getContext(), getChildFragmentManager(),
                 new SimplePagerAdapter.Page(R.string.es_menu_today, createFragmentProducer(TopicFeedType.EVERYONE_POPULAR_TODAY)),
                 new SimplePagerAdapter.Page(R.string.es_menu_week, createFragmentProducer(TopicFeedType.EVERYONE_POPULAR_THIS_WEEK)),
                 new SimplePagerAdapter.Page(R.string.es_menu_month, createFragmentProducer(TopicFeedType.EVERYONE_POPULAR_THIS_MONTH)),
@@ -75,10 +29,6 @@ public class PopularFeedFragmentTabs extends BaseFragment {
 
     private Producer<Fragment> createFragmentProducer(TopicFeedType topicFeedType) {
         return () -> PopularFeedFragmentTabs.createForFeedType(topicFeedType);
-    }
-
-    protected void onPageSelected(int position) {
-        // do nothing
     }
 
     public static PopularFeedFragment createForFeedType(TopicFeedType topicFeedType) {
